# Refactor string parsing to use std::from_chars for better performance