--- a/src/lib/es3n1n/common/string_parser.hpp
+++ b/src/lib/es3n1n/common/string_parser.hpp
@@ -3,16 +3,63 @@
 
 #include <any>
 #include <cassert>
+#include <charconv>
 #include <cstdint>
 #include <string>
 
-namespace string_parser {
+namespace string_parser2 {
+
+    namespace detail {
+
+        // \note: @annihilatorq: this scope is a mess, but idk how to make
+        // it more readable & cleaner, unfortunately, std::from_chars
+        // does not support strings with "0x" prefixes
+
+        /// Check if the string represents a negative value
+        ///
+        bool is_negative_value(const std::string_view view) {
+            return view.front() == '-';
+        }
+
+        /// Since std::from_chars doesn't support '0x' prefixed hex strings,
+        /// we need to process and remove prefixes before usage.
+        ///
+        void strip_hex_prefix(std::string_view& view, bool is_negative, int base) {
+            if (view.starts_with("-0x") || view.starts_with("-0X")) {
+                view.remove_prefix(3);
+            } else if (view.starts_with("0x") || view.starts_with("0X")) {
+                view.remove_prefix(2);
+            }
+        }
+
+        template <typename T>
+        [[nodiscard]] T parse_from_chars(std::string_view s, int base = 10) {
+            const auto is_negative = is_negative_value(s);
+            const auto is_hex = base == 16;
+
+            if (is_hex) {
+                strip_hex_prefix(s, is_negative, base);
+            }
+
+            T value;
+            auto [ptr, ec] = std::from_chars(s.data(), s.data() + s.size(), value, base);
+            if (ec == std::errc::invalid_argument) {
+                throw std::invalid_argument("Failed to parse integer from string: invalid argument");
+            } else if (ec == std::errc::result_out_of_range) {
+                throw std::out_of_range("Failed to parse integer from string: out of range");
+            }
+
+            /// Reintroduce HEX as negative if it was negative
+            return is_negative && is_hex ? static_cast<T>(-static_cast<std::make_signed_t<T>>(value)) : value;
+        }
+    } // namespace detail
+
     /// \brief Parse int32 from string
     /// \param s string that contain int32
     /// \param base base (10 for decimal, 16 for hex, etc)
     /// \return parsed value
     [[nodiscard]] inline std::int32_t parse_int32(const std::string_view s, const std::size_t base = 10) {
-        return std::stoi(s.data(), nullptr, static_cast<int>(base));
+        return detail::parse_from_chars<std::int32_t>(s, static_cast<int>(base));
     }
 
     /// \brief Parse uint32 from string
@@ -20,7 +67,7 @@ namespace string_parser {
     /// \param base base (10 for decimal, 16 for hex, etc)
     /// \return parsed value
     [[nodiscard]] inline std::uint32_t parse_uint32(const std::string_view s, const std::size_t base = 10) {
-        return std::stoul(s.data(), nullptr, static_cast<int>(base));
+        return detail::parse_from_chars<std::uint32_t>(s, static_cast<int>(base));
     }
 
     /// \brief Parse int8 from string
@@ -131,8 +178,8 @@ namespace string_parser {
         assert(ref.has_value());
         const auto hash = ref.type().hash_code();
 
-/// This is ugly, however, i can't calculate typeid(T).hash_code() in constexpr so
-/// no optimized switch cases :shrug: (could be implementation defined)
+        /// This is ugly, however, i can't calculate typeid(T).hash_code() in constexpr so
+        /// no optimized switch cases :shrug: (could be implementation defined)
 #define MAKE_CASE(type)                                   \
     if (hash == typeid(type).hash_code()) {               \
         return serialize<type>(std::any_cast<type>(ref)); \
@@ -146,6 +193,6 @@ namespace string_parser {
 
 #undef MAKE_CASE
 
-        throw std::runtime_error(std::format("parse_to_any: Unable to serialize -> unsupported type"));
+        throw std::runtime_error(std::format("serialize_any: Unable to serialize -> unsupported type"));
     }
-} // namespace string
+} // namespace string_parser2