--- a/src/lib/es3n1n/common/string_parser.hpp
+++ b/src/lib/es3n1n/common/string_parser.hpp
@@ -7,7 +7,7 @@
 #include <cstdint>
 #include <string>
 
-namespace string_parser2 {
+namespace string_parser {
 
     namespace detail {