--- a/src/tests/string_parser.cpp
+++ b/src/tests/string_parser.cpp
@@ -3,16 +3,24 @@
 
 TEST(string_parser, basics) {
     EXPECT_EQ(string_parser::parse_int32("-123"), -123);
+    EXPECT_EQ(string_parser::parse_int32("0x123", 16), 0x123);
+    EXPECT_EQ(string_parser::parse_int32("0X123", 16), 0x123);
+    EXPECT_EQ(string_parser::parse_int32("-0X123", 16), -0x123);
     EXPECT_EQ(string_parser::parse_int32("-0x123", 16), -0x123);
 
     EXPECT_EQ(string_parser::parse_uint32("123"), 123);
     EXPECT_EQ(string_parser::parse_uint32("0x123", 16), 0x123);
+    EXPECT_EQ(string_parser::parse_uint32("0X123", 16), 0x123);
 
     EXPECT_EQ(string_parser::parse_int8("-12"), -12);
+    EXPECT_EQ(string_parser::parse_int8("0x12", 16), 0x12);
+    EXPECT_EQ(string_parser::parse_int8("0X12", 16), 0x12);
+    EXPECT_EQ(string_parser::parse_int8("-0X12", 16), -0x12);
     EXPECT_EQ(string_parser::parse_int8("-0x12", 16), -0x12);
 
     EXPECT_EQ(string_parser::parse_uint8("12"), 12);
-    EXPECT_EQ(string_parser::parse_uint32("0x12", 16), 0x12);
+    EXPECT_EQ(string_parser::parse_uint8("0x12", 16), 0x12);
+    EXPECT_EQ(string_parser::parse_uint8("0X12", 16), 0x12);
 
     EXPECT_EQ(string_parser::parse_bool("true"), true);
     EXPECT_EQ(string_parser::parse_bool("1"), true);