## Summary

This Pull Request proposes a refactor of the string parsing functions to use `std::from_chars` instead of `std::stoi`, `std::stoul`, etc. This change significantly improves performance by leveraging the more efficient `std::from_chars` function, so i think it could be useful.

## Changes

- Refactored `parse_int32`, `parse_uint32`, `parse_int8`, `parse_uint8` to use `std::from_chars`.
- Added namespace `detail` to handle the conversion with appropriate error checking.
- Updated `parse` and `parse_to_any` functions to use the new parsing methods.

## Justification

- `std::from_chars` provides a more efficient and faster way to convert strings to integers.

## Tests

- Verified that all existing tests pass.

## Benchmark Results

The following benchmark results demonstrate the performance improvements of the new parsing functions compared to the current implementation:
```
Run on (12 X 2592 MHz CPU s)
CPU Caches:
  L1 Data 32 KiB (x6)
  L1 Instruction 32 KiB (x6)
  L2 Unified 256 KiB (x6)
  L3 Unified 12288 KiB (x1)
-----------------------------------------------------------------
Benchmark                       Time             CPU   Iterations
-----------------------------------------------------------------
BM_ParseInt32_CURRENT        45.0 ns         44.3 ns     14451613
BM_ParseInt32_NEW            7.80 ns         7.81 ns    112000000
BM_ParseUInt32_CURRENT       33.7 ns         33.8 ns     20363636
BM_ParseUInt32_NEW           7.43 ns         7.53 ns    112000000
BM_ParseInt8_CURRENT         29.8 ns         29.3 ns     22400000
BM_ParseInt8_NEW             6.14 ns         6.00 ns    112000000
BM_ParseUInt8_CURRENT        30.4 ns         30.0 ns     22400000
BM_ParseUInt8_NEW            6.14 ns         6.14 ns    112000000
BM_ParseBool_CURRENT        0.359 ns        0.359 ns   1000000000
BM_ParseBool_NEW            0.365 ns        0.359 ns   1000000000
```

## Notes

The refactor ensures that all integer parsing functions are now using `std::from_chars`, which is more efficient and modern. This should lead to better performance, especially in high-load scenarios.

Please let me know if any additional changes or improvements need to be made. Thank you for this useful and easy-to-use repository!